requirements.txt
accelerate==1.10.1
addict==2.4.0
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.113.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
kiwisolver==1.4.7
llm-rankers==0.0.2
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
oauthlib==3.3.1
openai==1.107.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.13.1
segment-anything==1.0
sentencepiece==0.2.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
supervision==0.26.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yapf==0.43.0
zipp==3.23.0
